Modify text-highlighter
