+++ src/inject/store/mutations.js
@@ -9,11 +9,12 @@ export default {
     state.canNextPage = payload.canNextPage
   },
   [types.FETCH_COMMENT_DATA] (state, payload) {
-    state.commentData = payload
+    state.commentData = payload.data
     state.loading = true
   },
   [types.SEARCH_COMMENT_DATA] (state, payload) {
-    state.currentCommentData = payload
+    state.currentCommentData = payload.data
+    state.searchText = payload.searchText
   },
   [types.PREV_PAGE] (state) {
   },
